BUG: bumping ref count on the master socket as opposed to slave
I was increasing the reference count on the master listen socket as
opposed to increasing the ref count on the slave.  With a decrease in the
ref count, the kernel deallocated the slave socket.  However, since the
slave socket had skbuffs attached to it, the socket could not be
deallocated.  The kernel left the socket alive and waited for the
skbuffs to be freed before freeing the socket itself.