BUG FIX: do not free an uninitialized multi-accept socket
The order of system calls to start a listening socket is:

 socket
 bind
 listen
 accept

To start a multi-accept socket the order is:

 socket
 setsockopt(TCP_MULTI_ACCEPT)
 bind
 listen
 accept

If bind fails, however, the kernel will close the socket. Since listen
was never called, the multi-accept state is not allocated. The old code
attempted to free state that had not been allocated yet.