Bug fix.  Prevent race between vma_adjust and page_add_new_anon_rmap.
Adding an anon_rmap needs to set the index of the new page, which is
computed based on address, vm_start, and vm_pgoff.  Thus, we need a
consistent view of these in page_add_new_anon_rmap, as well as in our
unmap race check.  Use a seqcount to accomplish this.