Bring back the cbtree library code, with lots of cleanup.
In particular, compiling it as a regular user program works again.  I
added a very small compatibility layer so that most of the code
doesn't have to care if it's user or kernel.  I rearranged things a
bit so that delay-freeing is no longer required for single threaded
user-space execution and got rid of dfree (we couldn't rely on this
trick for kernel RCU anyway).  Add some memory barriers to make RCU
happy.