use the local sk when putting thread on wait queue
When executing the poll function the first time around, the calling
thread is put on the socket's wait queue via sock_poll_wait. The best
policy is to place the thread asking to poll on the wait queue of the
local sk not on the wait queue of the first core.

This modification makes poll work: it is now possible for all threads to
poll on their local sk's wait queues.